backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
alembic
bcrypt
python-jose[cryptography]
python-dotenv